pyproject.toml
[tool.poetry]
name = "yandexid"
version = "0.3.0"
description = "Yandex ID oauth API wrapper"
authors = ["Maxim Mosin <max@mosin.pw>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/LulzLoL231/yandexid"
repository = "https://github.com/LulzLoL231/yandexid"
keywords = ["yandex", "oauth", "api", "wrapper", "yandexid", "yandex-id"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
packages = [
    { include = "yandexid", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.23.3"
pydantic = "^1.10.4"
PyJWT = "^2.6.0"

[tool.poetry.dev-dependencies]
mypy = "^0.991"
flake8 = "^6.0.0"
autopep8 = "^2.0.1"
bumpver = "^2022.1120"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.bumpver]
current_version = "0.3.0"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "Bump version {old_version} -> {new_version}"
commit = false
tag = false
push = false

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'current_version = "{version}"',
    'version = "{version}"',
]
"src/yandexid/__meta.py" = [
    "__version__ = '{version}'",
]
